% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_field}
\alias{get_field}
\title{get search options for field}
\usage{
get_field()
}
\value{
character vector로 분야:field를 확인합니다.
}
\description{
\code{\link{search_speech}} 함수에서 사용할 field 조건을 출력합니다.
분야에 대한 값들을 확인할 수 있습니다.
인터넷을 사용하는 기능입니다.
}
