% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speech.R
\name{get_speech}
\alias{get_speech}
\title{get speach}
\usage{
get_speech(target, paragraph = FALSE)
}
\arguments{
\item{paragraph}{문단 단위로 분리할지 말지를 정합니다. 기본값은 FALSE 입니다.
logical 을 입력으로 받습니다.}

\item{link}{검색어를 입력합니다. 기본값은 전체를 검색합니다.
character를 입력으로 받습니다.}
}
\value{
tibble 자료형으로 key, value 의 2개 컬럼을 가집니다.
}
\description{
\code{\link{search_speech}}에서 얻은 주소:link를 기준으로 연설문을 가져옵니다.
한번에 한 주소만 가져올 수 있습니다.
}
