% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{search speech list}
\usage{
search(keyword = "", top_n = 20, president = "", field = "",
  event = "")
}
\arguments{
\item{keyword}{is naver news link.}

\item{top_n}{is what you want to get from news. Defualt is all.}

\item{president}{is how many you want to try again if error. Default is 3.}

\item{field}{is wait time to try again. Default is rnorm(1).}

\item{event}{async crawling if it is TRUE. Default is FALSE.}
}
\value{
Get char vectors for list
}
\description{
search speech list
}
