% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spidx.R
\docType{data}
\name{spidx}
\alias{spidx}
\title{speech index data}
\format{tibble with columns
\describe{
\item{president}{대통령}
\item{field}{분야}
\item{event}{유형}
\item{title}{연설문 제목}
\item{date}{날짜}
\item{link}{연설문 링크}
}}
\source{
http://www.pa.go.kr/research/contents/speech/index.jsp
}
\usage{
spidx
}
\description{
speech index data
}
\examples{
  spidx
}
\keyword{datasets}
