% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speech.R
\name{get_speech}
\alias{get_speech}
\title{get speech}
\usage{
get_speech(link, paragraph = FALSE)
}
\arguments{
\item{link}{게시글의 url 주소입니다. 한 번에 한 주소만 가능합니다.
\url{http://www.pa.go.kr/research/contents/speech/index.jsp?spMode=view&catid=c_pa02062&artid=1310475}
와 같이 생겼습니다.
character를 입력으로 받습니다.}

\item{paragraph}{문단 단위로 분리할지 말지를 정합니다.
기본값은 FALSE 입니다.
logical 을 입력으로 받습니다.}
}
\value{
tibble 자료형으로 title, date, president, place, field, event, source, paragraph, content의 9개 컬럼을 가집니다.
}
\description{
\code{\link{search_speech}}에서 얻은 주소:link를 기준으로 연설문을 가져옵니다.
한번에 한 주소만 가져올 수 있습니다.
}
